Split Player into focused partial classes.
Player.cs had grown to a 515-line grab bag of state, movement, combat, &
cosmetics. Split it into four small partial-class files with no behavior
change: Player.cs keeps exports, signals, replicated properties, node
wiring, & thin per-frame dispatchers; Player.Movement.cs, Player.Combat.cs,
& Player.Cosmetics.cs hold the behavior. No public, exported, replicated,
RPC, signal, or node-path names changed, so replication & scenes are
untouched. Also factored two tiny helpers: FindAimedPlayer (shared
camera-ray query for crosshair tint & punch) & RestoreBaseColor (shared
spawn-armor/normal color restore).
